Address the review on the localization foundation
Counts that drive plural selection now cross into String(format:) as Int64
rather than Int. Int is already 64-bit on every device this app supports,
so nothing was broken, but varargs give the compiler nothing to check and
the %lld contract is better stated at the call site than assumed.

The settings footer read a hardcoded 0.1.0 while the project ships
MARKETING_VERSION 1.0, so the screen was naming a version that does not
exist. Read CFBundleShortVersionString instead; the displayed version now
follows whatever is actually packaged.

+++ SharedLedger/Domain/LedgerNotification.swift
@@ -199,7 +199,7 @@ struct LedgerSettlementReminder: Equatable, Sendable {
         case .both: key = .notificationBodySettlementReminderBoth
         }
         return key.string(
-            arguments: [groupName, bookName, outstandingTransferCount],
+            arguments: [groupName, bookName, Int64(outstandingTransferCount)],
             locale: locale
         )
     }
